# backend/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: Foraged goods marketplace API
  version: 0.1.0
  # "|" keeps the newlines - the description renders as Markdown in the UI.
  description: |
    A marketplace for foraged goods — berries, mushrooms, herbs. Sellers post
    listings, buyers browse, message them and place orders. Payments are
    simulated; no money changes hands.

    ## Authentication

    Two tokens, with opposite properties.

    The **access token** is a signed JWT sent on every request:

        Authorization: Bearer <access_token>

    It is verified without touching the database, which also means it cannot be
    revoked - so it is short-lived (15 minutes by default, `ACCESS_TOKEN_TTL`).

    The **refresh token** is an opaque value in an HttpOnly `refresh_token`
    cookie, set by signup and login and valid for 7 days. `POST /auth/refresh`
    exchanges it for a new access token and rotates it. Only its hash is
    stored, so a database leak cannot resume a session.

    When an access token expires, the 401 carries
    `WWW-Authenticate: Bearer error="invalid_token"` - the signal to refresh
    rather than to show a login screen. `GET /auth/me` returns the identity
    behind a token that is still valid, for restoring a session on page load.

    Public endpoints (browsing listings, search, images) work without one.

    ## Ids

    Every id is a **uuid v7**: the leading 48 bits are a millisecond timestamp,
...
    something to parse.

    ## Errors

...

...

...
servers:
...

...
security:
  - bearerAuth: []
  - apiKeyAuth: []

# "components" is the reusable-definitions section. Nothing in here appears in
# the docs on its own - it only shows up where something $refs it.
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A key from `POST /me/api-keys`, for machine clients. It acts as its
        owner - same permissions, same ownership checks - and is rate limited
        per key rather than per IP.

        Keys cannot manage keys: creating, listing and revoking need a session,
        so a leaked key cannot mint a replacement and survive its own
        revocation.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # documentation only; no tool verifies the format
      description: |
        Access token issued by signup or login. Expired and missing tokens are
        both answered with 401.
  schemas:
...
    Error:
      type: object
      required: [error] # "details" is omitempty in Go, so it is not required
      properties:
        error:
          type: string
          description: |
            Safe to show a user on 4xx, and written for display: sentence case,
            no trailing full stop. Render it as-is rather than transforming it.
            On 5xx this is always "Something went wrong" — the cause is in the
            server log, tagged with the `X-Request-ID` on the response.
          examples: ["Listing not found"]
        details:
          type: object
          # "any key, string value" - the YAML spelling of map[string]string.
          additionalProperties:
            type: string
          description: Reserved for field-level validation errors; absent today.
...
    CreateOrderInput:
      type: object
      required: [listing_id, quantity]
      properties:
        listing_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          exclusiveMinimum: 0
    Order:
      type: object
      required:
        [
          id,
          listing_id,
          listing_title,
          buyer_id,
          seller_id,
          quantity,
          unit_price,
          total_price,
          status,
          seller_handed_over_at,
          buyer_received_at,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        listing_id:
          type: string
          format: uuid
        listing_title:
          type: string
          description: |
            Snapshotted when the order was placed, so the order still reads
            correctly after the listing is edited.
        buyer_id:
          type: string
          format: uuid
        seller_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
        unit_price:
          type: number
          format: double
          description: |
            Price per unit, snapshotted when the order was placed, so the order
            still reads correctly after the listing is repriced.

            Stored as `NUMERIC(10,2)` and converted only on the way out —
            `total_price` is multiplied by Postgres in exact decimal, never in
            Go. Trailing zeroes are not preserved: 18.50 is sent as 18.5, and
            74.00 as 74. Format for display client-side.
          examples: [18.5]
        total_price:
          type: number
          format: double
          description: unit_price × quantity, computed in the database.
          examples: [74]
        status:
          $ref: "#/components/schemas/OrderStatus"
        seller_handed_over_at:
          # "null OR a timestamp" - the Go field is *time.Time.
          type: [string, "null"]
          format: date-time
          description: Set when the seller marked handover; null until then.
        buyer_received_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserSettings:
      type: object
      required: [show_online_status]
      properties:
        show_online_status:
          type: boolean
    UpdateSettingsInput:
      type: object
      # Required because the Go field is a *bool specifically to tell
      # "false" apart from "not sent" - and "not sent" is a 400.
      required: [show_online_status]
      properties:
        show_online_status:
          type: boolean
...
    ProfileUpdate:
      type: object
      description: |
        Every field is optional and each accepts three states, so a PATCH can
        say "leave this alone", "clear it", or "set it" per field:

        | Sent | Effect |
        |---|---|
        | key absent | keep the current value |
        | `null` **or** `""` | clear the field to `null` |
        | a value | replace it (surrounding whitespace is trimmed) |

        Note `null` and `""` behave the same. RFC 7396 merge-patch would treat
        `""` as a value; this API treats an empty string as "clear", because a
        form that submits a cleared input sends `""` rather than `null`.

        Unknown fields are rejected with a 400 - with every field optional, a
        typo is otherwise indistinguishable from an omission.
      additionalProperties: false
      properties:
        firstname:
          type: [string, "null"]
          maxLength: 150
        lastname:
          type: [string, "null"]
          maxLength: 150
        bio:
          type: [string, "null"]
          maxLength: 1000
        phone_number:
          type: [string, "null"]
          maxLength: 15
        date_of_birth:
          type: [string, "null"]
          format: date
          description: |
            `YYYY-MM-DD`. Rejected if it is not a plain date, is in the future,
            or falls before 1900.
          examples: ["2001-05-14"]
        location:
          type: [string, "null"]
          maxLength: 100
          description: |
            Free text, stored on the user's address rather than the profile.
            This is what the `location` filter on `/listings/search` matches.
          examples: ["Espoo"]
    OwnProfile:
      type: object
      description: |
        Your own profile. Unset fields are `null` rather than absent, so a form
        can render an empty input without deciding whether the key exists.
      required:
        [id, username, email, firstname, lastname, bio, phone_number, date_of_birth, location]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        firstname:
          type: [string, "null"]
        lastname:
          type: [string, "null"]
        bio:
          type: [string, "null"]
        phone_number:
          type: [string, "null"]
        date_of_birth:
          type: [string, "null"]
          format: date
          examples: ["2001-05-14"]
        location:
          type: [string, "null"]
    PublicProfile:
      type: object
      description: |
        Somebody else's profile. `email`, `phone_number` and `date_of_birth`
        are not blanked here - the shape has no such fields, so they cannot be
        exposed by an oversight.
      required: [id, username, firstname, lastname, bio, location, presence]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstname:
          type: [string, "null"]
        lastname:
          type: [string, "null"]
        bio:
          type: [string, "null"]
        location:
          type: [string, "null"]
        presence:
          $ref: "#/components/schemas/Presence"

  # Whole reusable responses, not just schemas - status code descriptions we
  # would otherwise retype on all 35 operations.
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json: # the media type the body is sent as
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: |
        No token, or one that is expired or invalid. The body is the same in
        every case; `WWW-Authenticate` says which, because a client's response
        differs: refresh, or log in.
      headers:
        WWW-Authenticate:
          description: |
            `Bearer error="invalid_token", error_description="The access token expired"`
            when the token was valid but has aged out - retry through
            `POST /auth/refresh`. A bare `Bearer` for a missing, malformed or
            forged token - the client should log in. A forged token reads as
            absent on purpose.
          schema:
            type: string
          examples:
            expired:
              value: Bearer error="invalid_token", error_description="The access token expired"
            absent:
              value: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated, but not permitted — the resource belongs to someone
        else, or the account does not hold the role the route requires.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        No such resource — also returned when a resource exists but the caller
        is not a participant, so its existence is not revealed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: |
        Too many requests on this API key. Only key-authenticated requests are
        limited; browser sessions are not.
      headers:
        Retry-After:
          description: Seconds until a request will be accepted again.
          schema:
            type: integer
            examples: [2]
        X-RateLimit-Limit:
          description: |
            The bucket's capacity. Sent on every key-authenticated response,
            not only on a refusal.
          schema:
            type: integer
            examples: [60]
        X-RateLimit-Remaining:
          description: Requests left in the current bucket.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Something failed on our side; the cause is in the log.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# Each key is a URL relative to the server above, so "/listings" is really
# http://localhost:8080/api/v1/listings.
paths:
...
  /me/profile:
    get:
      tags: [me]
      summary: Get your own profile
      operationId: getOwnProfile
      responses:
        "200":
          description: Your profile, including the fields only you can see.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: |
            The account behind this token no longer exists. Signup creates the
            profile row, so a missing profile means a missing user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    patch:
      tags: [me]
      summary: Update your own profile
      operationId: updateOwnProfile
      description: |
        A partial update: fields you do not mention keep their current value.
        Both halves are written in one transaction, since `location` lives on
        the address rather than the profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: The profile after the change.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnProfile"
        "400":
          description: |
            A malformed body, an unknown field, a value over its limit, a
            string containing a null byte, or a `date_of_birth` that is not a
            plain past date after 1900.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [me]
      summary: Get someone else's public profile
      operationId: getPublicProfile
      security: []
      description: |
        Public, like the listing reads - a seller's profile should be readable
        while browsing. `presence` honours that user's `show_online_status`:
        with it off, the response is `{"is_online": false}` with no
        `last_seen_at`, identical to a user who has never been seen.
      responses:
        "200":
          description: The public view of that profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "400":
          description: The id is not a UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /orders:
    post:
      tags: [orders]
      summary: Place an order
      operationId: createOrder
      description: |
        Decrements the listing's stock in the same transaction, so two buyers
        racing for the last item cannot both succeed. Starts as `pending`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order placed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: |
            Bad body, non-positive quantity, or your own listing — you cannot
            buy from yourself.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not enough stock available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    get:
      tags: [orders]
      summary: List your orders
      operationId: getOrders
      description: Every order where you are the buyer OR the seller.
      responses:
        "200":
          description: Your orders.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [orders]
      summary: Get one order
      operationId: getOrder
      responses:
        "200":
          description: The order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/events:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [orders]
      summary: An order's history
      operationId: getOrderEvents
      description: |
        Every transition the order made, oldest first: who did what, and when.
        A completed order has five entries - placed, confirmed, both halves of
        the handshake, and the completion.

        The two handshake entries share a `created_at`, because they are
        written in one transaction; they are ordered by `id` behind it.

        Buyer and seller only.
      responses:
        "200":
          description: The timeline, oldest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            You are not the buyer or the seller. Note orders answer this with
            403 where conversations answer 404 - so an outsider can tell that
            an order id exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/confirm:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      summary: Seller accepts the order
      operationId: confirmOrder
      description: "`pending` → `confirmed`. Seller only."
      responses:
        "200":
          description: The order, now `confirmed`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The order is not `pending`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/handover:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      summary: Seller marks the goods handed over
      operationId: handoverOrder
      description: |
        Half of the two-sided handshake. Stamps `seller_handed_over_at`; the
        status only becomes `completed` once the buyer has called `/receive`
        as well. Calling it twice is a 409.
      responses:
        "200":
          description: |
            The order — `completed` if the buyer already confirmed receipt,
            otherwise still `confirmed` with the timestamp now set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Wrong status, or you already marked handover.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/receive:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      summary: Buyer confirms receipt
      operationId: receiveOrder
      description: |
        The other half of the handshake. Stamps `buyer_received_at`; the
        status becomes `completed` only once the seller has handed over too.
      responses:
        "200":
          description: The order, `completed` if both sides have now marked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Wrong status, or you already confirmed receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /orders/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      summary: Cancel an order
      operationId: cancelOrder
      description: |
        Either side, while the order is `pending` or `confirmed`. The reserved
        stock goes back to the listing. Once either side has marked the
        handshake, cancelling is refused.
      responses:
        "200":
          description: The order, now `cancelled`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Already finished, or handover has started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"

  # ---------------------------------------------------------------------------
  # chat
  # ---------------------------------------------------------------------------
...
  /me/settings:
    get:
      tags: [me]
      summary: Get your settings
      operationId: getSettings
      responses:
        "200":
          description: Your settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    patch:
      tags: [me]
      summary: Update your settings
      operationId: updateSettings
      description: |
        Turning `show_online_status` off makes you look permanently offline to
        everyone else — your presence stops being reported, not just hidden.
        Omitting the field entirely is a 400, not a no-op.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsInput"
      responses:
        "200":
          description: The settings after the change.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /me/unread:
    get:
      tags: [me, chat]
      summary: Your total unread message count
      operationId: getUnreadCount
      description: Across every conversation — for a badge in the header.
      responses:
        "200":
          description: The count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCount"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"